Add an API.
