+++ api.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const findUnusedVars = require('./src/findUnusedVars');
+
+module.export = {
+    find(folder) {
+        return findUnusedVars(folder, false);
+    }
+};
