+++ package.json
@@ -4,7 +4,11 @@
   "description": "A small tool to find unused Sass variables",
   "author": "XhmikosR <xhmikosr@gmail.com>",
   "license": "MIT",
-  "main": "index.js",
+  "main": "api.js",
+  "bin": {
+    "find-unused-sass-variables" : "./index.js",
+    "fusv" : "./index.js"
+  },
   "scripts": {
     "eslint": "eslint .",
     "lint": "npm run eslint",
@@ -25,6 +29,8 @@
     "eslint": "^4.12.0"
   },
   "files": [
+    "src/findUnusedVars.js",
+    "api.js",
     "index.js"
   ],
   "engines": {
